refactors OverallStatPieChart
